Allow changing log levels of different loggers independently
As we add more logging, it will be useful to split it into different
loggers, so that it's possible to only enable the verbose logging of one
particular component of Theia.  This patch allows the user to select the
log level for each logger indepedently.

First, it adds names to loggers, so that the user can be presented a
list of existing loggers.  The loggers are uniquely identified using
these names.  This allows the frontend and backend to both create
loggers named "task", for example, which will refer in the end to the
same logger.  If the user changes the log level for the logger "task",
it will affect both the frontend and the backend.  Using names as the
unique identifier of the loggers makes the numerical id obsolete, so it
is removed.

The UI to allow changing the log level is based on the quick open menu
and invoked using a new "Set Log Level" command available in the command
palette.

Changing the level of the root logger changes the level of all loggers.

Signed-off-by: Simon Marchi <simon.marchi@polymtl.ca>